requirements.txt
python-decouple
elevate
typer
colorama
shellingham
pytest
requests
